PB-1891: Resolve cyclic import
