+++ service_alti.py
@@ -1,3 +1,3 @@
 """Service launcher to use Flask without wsgi.py
 """
-from app import app  # pylint: disable=unused-import
+from app.app import app  # pylint: disable=unused-import
