+++ wsgi.py
@@ -2,7 +2,7 @@ import multiprocessing
 
 from gunicorn.app.base import BaseApplication
 
-from app import app as application
+from app.app import app as application
 from app.helpers import get_logging_cfg
 from app.settings import ALTI_WORKERS
 from app.settings import GUNICORN_WORKER_TMP_DIR
